Move translationUtils to own file
Refs #397.
