+++ app/utils/translationUtils.js
@@ -1,0 +1,10 @@
+function getProperty(item, property, language = 'fi') {
+  if (item && item[property] && item[property][language]) {
+    return item[property][language];
+  }
+  return '';
+}
+
+export {
+  getProperty,
+};
